iasted/backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.10.0
pydantic-settings==2.6.0

# WebSocket
websockets==13.1
python-socketio==5.11.4

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.0
redis-om==0.3.2

# RabbitMQ & Celery
celery[redis]==5.4.0
kombu==5.4.2
amqp==5.3.1

# AI Services - STT
deepgram-sdk==3.8.3

# AI Services - LLM
openai==1.55.3
anthropic==0.39.0
google-generativeai==0.8.3

# AI Services - TTS
google-cloud-texttospeech==2.18.0
elevenlabs==1.12.0

# NLU & Embeddings
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.5.1

# LLM Orchestration
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pycasbin==1.37.0
python-dotenv==1.0.1
cryptography==44.0.0

# AWS
boto3==1.35.65
botocore==1.35.65

# Document Generation
weasyprint==63.1
jinja2==3.1.4
pypdf==5.1.0

# Monitoring & Logging
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==3.2.1
sentry-sdk[fastapi]==2.18.0

# HTTP Client
httpx==0.27.2
aiohttp==3.11.7

# Utils
python-dateutil==2.9.0.post0
pytz==2024.2
pydantic-extra-types==2.10.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.2

# Dev Tools
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2
pre-commit==4.0.1

